adding macros. to define a macro: macro foo(arg)  ... end
to use a macro:

@foo arg

for now, macros take just 1 argument. the macro is called on the
literal syntax of its argument, and its result is used in place of the
macro call.

while i'm at it, improving the internal interface to eval, and making
the front end a bit more efficient (it no longer has to wrap
everything in a thunk)

making `x` parse to @cmd "x", but with different escaping behavior as
discussed.