major update to modules:
replace "import M.*" with "using M".

import paths are relative to the current module, and search in parent modules
as necessary.

we are now stricter about name conflicts: once use of a name in a module is
resolved, the binding it refers to cannot change. currently the messages are
warnings, but can be changed to errors trivially.

exporting a name is not sufficient to "resolve" it.

fixes the following issues:

status:
base/ has been ported, sysimg builds, a few tests pass.
extras/ not yet touched.
a few more imports are probably needed in Base to handle methods added by
submodules.