Don't call parse-partial-sexp to find the last opening paren.
This doesn't work because Julia's character syntax cannot simply be
expressed with an Emacs syntax table. However, parse-partial-sexp
is depending on the syntax table to avoid strings and characters.

Instead, just move backwards until we find a paren that isn't in a
string, character or comment.

Fixes https://github.com/JuliaLang/julia/issues/8924