julia-mode: Indent manually inside strings with \n
Don't try to automatically indent when inside strings with newline
characters. We don't want to add indentation where it is not intended,
so any automatic formatting should not assume that we want to indent
inside a string. This also fixes #12419, because we won't get confused
by keywords in multi-line strings.

Because it is still useful, make TAB and <backtab> increase and decrease
indentation when called inside a string.