Remove font-lock-type-face for types at point of use
In the elisp manual, font-lock-type-face is described as being "for
the names of user-defined data types.". It is intended to font-lock
the point of type definition rather than use.

By using this font-lock on a list of builtin types, a visual
distinction is created between builtin and library/user-defined types
when none exists in reality.