Font lock tweaks (#102)
* Remove unused type-parameter regex

This regex was designed to operate on type parameters (e.g. Vector{T}
has T as a type parameter) to fontify them as types. However type
parameters can be values and not just types (e.g. NTuple{2,Int}) so
fontifying them as such is not correct.

* Remove font-lock-type-face for types at point of use

In the elisp manual, font-lock-type-face is described as being "for
the names of user-defined data types.". It is intended to font-lock
the point of type definition rather than use.

By using this font-lock on a list of builtin types, a visual
distinction is created between builtin and library/user-defined types
when none exists in reality.

* font-lock quoted symbols as constants instead of preprocessor

This fits better with a quoted symbols purpose in source-code. You
could also consider font-lock-string-face.

Closes #4.

* Remove "in" from font-lock keywords

When in a for-loop or comprehension, "in" is already
font-lock-keyword-face because of julia-forloop-in-regex.

* Remove unused font-locking for builtins

In the future, we could consider using font-lock-builtin-face for
everything returned by names(Base), but given that none of these are
truly builtin (can be overridden in a baremodule), I think it's best
to not use this face.

* Add missing to font-lock-constant-face

* Use font-lock-keyword face for ternary ? and :

* Add baremodule to block start

* Use terser setq-local in place of make-local-variable

This is equivalent to the current implementation with
make-local-variable, but setq-local is only available in emacs 24.3+.

* Remove regex term matching obsolete parametric function spelling

Pre Julia 1.0, parametric functions could be spelled as f{T}(args...)
but now the only correct spelling is f(args...) where T.

* Use font-lock-keyword-face for = and ∈ in loops

Previously, only ∈ was font-locked and only if it had whitespace
around it. This new regex correctly font-locks "for i=1:10" as well as
"for i = 1:10". This commit also adds tests for whether
font-lock-keyword-face is being correctly applied in for loops.

* Apply reviewer comments from PR 102

See also the previous commit for addition of forloop font-lock tests.