Add integration with project.el
This code is already used by eglot-jl to locate Julia projects and
will be useful in the future for other Julia packages (such as for
REPLs to decide which files they're associated with). Putting it in
the julia-mode package means it doesn't have to be redefined each time
it's used.