General refactor and improvements
These are changes unrelated to #15 that just make the code a bit more
readable. This commit should be reviewed separately.

Since the julia group was declared with a :prefix of julia- all of
these customization variables will be in the julia group anyway. Also,
since julia-mode didn't declare it's group, an anonymous julia-mode
group was getting created for faces.