Indent lines after hanging operator even if previous line contains # (#115)
* General refactor and improvements

These are changes unrelated to #15 that just make the code a bit more
readable. This commit should be reviewed separately.

Since the julia group was declared with a :prefix of julia- all of
these customization variables will be in the julia group anyway. Also,
since julia-mode didn't declare it's group, an anonymous julia-mode
group was getting created for faces.

* Move around definitions to get code byte-compiling cleanly

* Indent lines after hanging operator even if previous line contains #

Because the previous implementation used a regex to detect whether the
hanging operator was in a comment or not, it did not detect hanging
operators following a # in a string.

Fixes #19

* Add comment explaining automatic adding to latest defgroup