Add predicate to "around" latex symbol completion
This predicate ensures that the symbol you're trying to complete is a substring of the
completion candidates. With "before" completion this isn't necessary since there isn't
potential for the symbol you're trying to complete containing text that isn't intended to be
completed.

This commit also includes some refactors for clarity and DRY.