Strip whitespace in latexsub :exit-function
According to the docstring for completion-extra-properties, the "STRING" passed to the
:exit-function should be the bare text to which the field was completed, but helm-mode adds
an extra space at the end of "STRING". Since none of our latexsubs include whitespace, we
can safely strip whitespace in the :exit-function and work around this helm bug.

I will file an upstring bug report against helm, but I think it's worth fixing quickly here
since it's only a simple 3-line change.

Closes #204.